Extract try_make_solid into periodic_shell
Add a centralized try_make_solid helper in microgen.shape.periodic_shell that upgrades sewn shells (or compounds of shells) into TopoDS_Solid(s), reorients enclosed volume, and fuses multiple solids when needed. Export the helper via __all__ and update Tpms and Spinodoid to reuse it (removing duplicated implementations). Also tidy related imports (remove now-unused fuse_shapes import in tpms). The helper uses OCP APIs and is shared by TPMS and spinodoid shape creation.
